+++ mono.engine/bake/src/GraphDocument.cpp
@@ -93,13 +93,13 @@ namespace engine::bake {
 				break;
 			case OperationKind::AddFlipbook:
 				if (operation.Side > std::numeric_limits<uint8_t>::max() ||
-					operation.Frames > std::numeric_limits<uint8_t>::max()) {
+					operation.Frames > std::numeric_limits<uint16_t>::max()) {
 					offender = describe();
 					return DocumentStatus::Refused;
 				}
 				added = graph.AddFlipbook(
 					static_cast<uint8_t>(operation.Side),
-					static_cast<uint8_t>(operation.Frames),
+					static_cast<uint16_t>(operation.Frames),
 					operation.Number
 				);
 				break;
